requirements.txt
apache-airflow
python-dotenv
pandas
psycopg2-binary